Skeleton for the Document Upload Widget. Currently has the "upload" button and allows for choosing a file.
Thinking about how to display this: In order to guarentee that we'd be
able to render in the notification it'd have to be an img and not a pdf
because different browsers handle PDFs differently.
This is unfortunate. Will ask Luke about this. There's probably also an
iframe plan but uuuuughhhhh. Imgs wouldn't be awful...

Thinking about the UX here, I like the idea that the widget could store
the upload in localstorage and then the preview would load it. This
wouldn't work for PDFs though. The other issue is that it's tough for
browser security reasons to get the img full url path from the upload
button. Maybe I could pass that into the form somehow?

Internet recommends pushing the img to a server and loading it from
there but I don't wanna POST anything until we know the admin is settled
on their final upload or we'd have lots of draft-type uploads in the DB